LayoutTests/platform/mac/accessibility/js-accessibility.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<link rel="stylesheet" href="../../../fast/js/resources/js-test-style.css">
<script>
var successfullyParsed = false;
</script>
<script src="../../../fast/js/resources/js-test-pre.js"></script>
</head>
<body id="body">

<p id="description"></p>
<div id="console"></div>

<script>

    description("This tests the javascript interface to accessibility");

    var body = 0;
    var name = 0;
    function notifyCallback(notification) {  
        name = notification;
        if (name == "AXScreenChanged") {
            shouldBe("name", "'AXScreenChanged'");
        }
        else if (name == "AXElementsChanged") {
            shouldBe("name", "'AXElementsChanged'");
            body.removeNotificationListener();
            window.layoutTestController.notifyDone();
        }
    }
</script>

<script>
    var ax = window.navigator.accessibility;

    // screen reader should not be running within layout tests.
    shouldBe("ax.screenReader.active", "false");

    // should not cause an error.
    var version = ax.screenReader.version;

    // should not cause an error
    var activeElement = ax.screenReader.activeElement;

    if (window.accessibilityController) {
        window.layoutTestController.waitUntilDone();

        document.getElementById("body").focus();

        body = window.accessibilityController.focusedElement;

        var addedNotification = body.addNotificationListener(notifyCallback);
        shouldBe("addedNotification", "true");

        // should cause a AXScreenChanged to be sent out.
        ax.screenChanged();

        // should cause a AXElementsChanged to be sent out.
        ax.elementsChanged();
    }

    successfullyParsed = true;
</script>

<script src="../../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
