LayoutTests/platform/mac/accessibility/js-accessibility-expected.txt

This tests the javascript interface to accessibility

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ax.screenReader.active is false
PASS addedNotification is true
PASS successfullyParsed is true

TEST COMPLETE
PASS name is 'AXScreenChanged'
PASS name is 'AXElementsChanged'
